fix(ESSNTL-4142): get rid of circular deps to avoid early access registery context
