+++ src/AppEntry.js
@@ -1,7 +1,8 @@
 import React, { useMemo } from 'react';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { Provider } from 'react-redux';
-import { getStore, RegistryContext, updateReducers } from './store';
+import { getStore, updateReducers } from './store';
+import RegistryContext from './store/registeryContext';
 import App from './App';
 import { getBaseName } from '@redhat-cloud-services/frontend-components-utilities/helpers';
 import logger from 'redux-logger';
