+++ src/components/InventoryTable/InventoryTable.test.js
@@ -10,22 +10,8 @@ import { createPromise as promiseMiddleware } from 'redux-promise-middleware';
 import { BrowserRouter as Router } from 'react-router-dom';
 import toJson from 'enzyme-to-json';
 import { ConditionalFilter } from '@redhat-cloud-services/frontend-components/ConditionalFilter';
-import * as actions from '../../Utilities/constants';
+import * as loadSystems from '../../Utilities/sharedFunctions';
 import { mockSystemProfile } from '../../__mocks__/hostApi';
-
-jest.mock('../../store/actions', () => {
-    const actions = jest.requireActual('../../store/actions');
-    const { ACTION_TYPES } = jest.requireActual('../../store/action-types');
-    return {
-        __esModule: true,
-        ...actions,
-        loadEntities: () => ({
-            type: ACTION_TYPES.LOAD_ENTITIES,
-            payload: () => Promise.resolve({}),
-            meta: { showTags: undefined }
-        })
-    };
-});
 
 describe('InventoryTable', () => {
     let initialState;
@@ -50,7 +36,8 @@ describe('InventoryTable', () => {
                 }
             }
         };
-        spy = jest.spyOn(actions, 'loadSystems').mockImplementation(() => ({ type: 'reload' }));
+
+        spy = jest.spyOn(loadSystems, 'loadSystems').mockImplementation(() => ({ type: 'reload' }));
         mockSystemProfile.onGet().reply(200, { results: [] });
     });
