+++ src/components/SystemDetails/Ros.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { useRouteMatch } from 'react-router-dom';
-import { RegistryContext } from '../../store';
+import RegistryContext from '../../store/registeryContext';
 import AsyncComponent from '@redhat-cloud-services/frontend-components/AsyncComponent';
 
 const RosTab = () => {
@@ -10,8 +10,8 @@ const RosTab = () => {
     return <AsyncComponent
         appName="ros"
         module="./SystemDetail"
-        getRegistry={ getRegistry }
-        inventoryId={ params.inventoryId }
+        getRegistry={getRegistry}
+        inventoryId={params.inventoryId}
     />;
 };
