+++ src/components/SystemDetails/Vulnerability.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
 import { useRouteMatch } from 'react-router-dom';
-import { RegistryContext } from '../../store';
+import RegistryContext from '../../store/registeryContext';
 import AsyncComponent from '@redhat-cloud-services/frontend-components/AsyncComponent';
 
 const VulnerabilityTab = () => {
@@ -12,7 +12,7 @@ const VulnerabilityTab = () => {
         module="./SystemDetail"
         getRegistry={getRegistry}
         customItnlProvider
-        entity={ { id: params.inventoryId } }
+        entity={{ id: params.inventoryId }}
     />;
 };
