+++ src/store/registeryContext.js
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+const RegistryContext = createContext({
+    getRegistry: () => { }
+});
+
+export default RegistryContext;
